src/api-service/requirements.txt
streamlit
yfinance
matplotlib
phidata
openai
plotly
seaborn
numpy
uvicorn
fastapi
requests
kaleido
